split rx and tx sync streamers into separate files.
