+++ src/streamers/mod.rs
@@ -1,22 +1,13 @@
-use std::borrow::Borrow;
-use std::marker::PhantomData;
-use std::sync::Arc;
 use std::time::Duration;
 
-use libbladerf_sys as sys;
-
-use crate::BladeRF;
-use crate::BladeRf1;
-use crate::BladeRf2;
-use crate::BladeRfAny;
-use crate::Channel;
-use crate::ChannelLayout;
-use crate::ChannelLayoutRx;
-use crate::ChannelLayoutTx;
 use crate::Error;
 use crate::Result;
-use crate::RxChannel;
-use crate::SampleFormat;
+
+mod rx_sync_stream;
+pub use rx_sync_stream::*;
+
+mod tx_sync_stream;
+pub use tx_sync_stream::*;
 
 pub struct SyncConfig {
     pub(crate) num_buffers: u32,
@@ -69,358 +60,5 @@ impl Default for SyncConfig {
             num_transfers: 8,
             stream_timeout: 3500,
         }
-    }
-}
-
-pub struct RxSyncStream<T: Borrow<D>, F: SampleFormat, D: BladeRF> {
-    pub(crate) dev: T,
-    pub(crate) layout: ChannelLayoutRx,
-    pub(crate) _devtype: PhantomData<D>,
-    pub(crate) _format: PhantomData<F>,
-}
-
-impl<T: Borrow<D>, F: SampleFormat, D: BladeRF> RxSyncStream<T, F, D> {
-    pub fn read(&self, buffer: &mut [F], timeout: Duration) -> Result<()> {
-        let res = unsafe {
-            sys::bladerf_sync_rx(
-                self.dev.borrow().get_device_ptr(),
-                buffer.as_mut_ptr() as *mut _,
-                buffer.len() as u32,
-                std::ptr::null_mut(),
-                timeout.as_millis() as u32,
-            )
-        };
-        check_res!(res);
-        Ok(())
-    }
-
-    fn new(dev: T, config: &SyncConfig, layout: ChannelLayoutRx) -> Result<RxSyncStream<T, F, D>> {
-        unsafe {
-            dev.borrow().set_sync_config::<F>(config, layout.into())?;
-        }
-
-        Ok(RxSyncStream {
-            dev,
-            layout,
-            _devtype: PhantomData,
-            _format: PhantomData,
-        })
-    }
-}
-
-impl<'a, F: SampleFormat, D: BladeRF> RxSyncStream<&'a D, F, D> {
-    fn reconfigure_inner<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-        layout: ChannelLayoutRx,
-    ) -> Result<RxSyncStream<&'a D, NF, D>> {
-        RxSyncStream::new(self.dev, config, layout)
-    }
-}
-
-impl<F: SampleFormat, D: BladeRF> RxSyncStream<Arc<D>, F, D> {
-    fn reconfigure_inner<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-        layout: ChannelLayoutRx,
-    ) -> Result<RxSyncStream<Arc<D>, NF, D>> {
-        RxSyncStream::new(self.dev.clone(), config, layout)
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// RX Stream Brf1
-
-impl<T: Borrow<BladeRf1> + Clone, F: SampleFormat> RxSyncStream<T, F, BladeRf1> {
-    pub fn enable(&self) -> Result<()> {
-        self.dev.borrow().set_enable_module(Channel::Rx0, true)
-    }
-
-    pub fn disable(&self) -> Result<()> {
-        self.dev.borrow().set_enable_module(Channel::Rx0, false)
-    }
-}
-
-impl<'a, F: SampleFormat> RxSyncStream<&'a BladeRf1, F, BladeRf1> {
-    pub fn reconfigure<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-    ) -> Result<RxSyncStream<&'a BladeRf1, NF, BladeRf1>> {
-        self.reconfigure_inner(config, ChannelLayoutRx::SISO(RxChannel::Rx0))
-    }
-}
-
-impl<F: SampleFormat> RxSyncStream<Arc<BladeRf1>, F, BladeRf1> {
-    pub fn reconfigure<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-    ) -> Result<RxSyncStream<Arc<BladeRf1>, NF, BladeRf1>> {
-        self.reconfigure_inner(config, ChannelLayoutRx::SISO(RxChannel::Rx0))
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// RX Stream Brf2
-
-impl<T: Borrow<BladeRf2> + Clone, F: SampleFormat> RxSyncStream<T, F, BladeRf2> {
-    pub fn enable(&self) -> Result<()> {
-        match self.layout {
-            ChannelLayoutRx::SISO(ch) => self.dev.borrow().set_enable_module(ch.into(), true),
-            ChannelLayoutRx::MIMO => {
-                self.dev.borrow().set_enable_module(Channel::Rx0, true)?;
-                self.dev.borrow().set_enable_module(Channel::Rx1, true)?;
-                Ok(())
-            }
-        }
-    }
-
-    pub fn disable(&self) -> Result<()> {
-        match self.layout {
-            ChannelLayoutRx::SISO(ch) => self.dev.borrow().set_enable_module(ch.into(), false),
-            ChannelLayoutRx::MIMO => {
-                self.dev.borrow().set_enable_module(Channel::Rx0, false)?;
-                self.dev.borrow().set_enable_module(Channel::Rx1, false)?;
-                Ok(())
-            }
-        }
-    }
-}
-
-impl<'a, F: SampleFormat> RxSyncStream<&'a BladeRf2, F, BladeRf2> {
-    pub fn reconfigure<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-        layout: ChannelLayoutRx,
-    ) -> Result<RxSyncStream<&'a BladeRf2, NF, BladeRf2>> {
-        self.reconfigure_inner(config, layout)
-    }
-}
-
-impl<F: SampleFormat> RxSyncStream<Arc<BladeRf2>, F, BladeRf2> {
-    pub fn reconfigure<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-        layout: ChannelLayoutRx,
-    ) -> Result<RxSyncStream<Arc<BladeRf2>, NF, BladeRf2>> {
-        self.reconfigure_inner(config, layout)
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// RX Stream BrfAny
-
-impl<T: Borrow<BladeRfAny> + Clone, F: SampleFormat> RxSyncStream<T, F, BladeRfAny> {
-    pub fn enable(&self) -> Result<()> {
-        match self.layout {
-            ChannelLayoutRx::SISO(ch) => self.dev.borrow().set_enable_module(ch.into(), true),
-            ChannelLayoutRx::MIMO => {
-                self.dev.borrow().set_enable_module(Channel::Rx0, true)?;
-                self.dev.borrow().set_enable_module(Channel::Rx1, true)?;
-                Ok(())
-            }
-        }
-    }
-
-    pub fn disable(&self) -> Result<()> {
-        match self.layout {
-            ChannelLayoutRx::SISO(ch) => self.dev.borrow().set_enable_module(ch.into(), false),
-            ChannelLayoutRx::MIMO => {
-                self.dev.borrow().set_enable_module(Channel::Rx0, false)?;
-                self.dev.borrow().set_enable_module(Channel::Rx1, false)?;
-                Ok(())
-            }
-        }
-    }
-}
-
-impl<'a, F: SampleFormat> RxSyncStream<&'a BladeRfAny, F, BladeRfAny> {
-    pub fn reconfigure<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-        layout: ChannelLayoutRx,
-    ) -> Result<RxSyncStream<&'a BladeRfAny, NF, BladeRfAny>> {
-        self.reconfigure_inner(config, layout)
-    }
-}
-
-impl<F: SampleFormat> RxSyncStream<Arc<BladeRfAny>, F, BladeRfAny> {
-    pub fn reconfigure<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-        layout: ChannelLayoutRx,
-    ) -> Result<RxSyncStream<Arc<BladeRfAny>, NF, BladeRfAny>> {
-        self.reconfigure_inner(config, layout)
-    }
-}
-
-impl<T: Borrow<D>, F: SampleFormat, D: BladeRF> Drop for RxSyncStream<T, F, D> {
-    fn drop(&mut self) {
-        // Ignore the results, just try disable both channels even if they don't exist on the dev.
-        let _ = self.dev.borrow().set_enable_module(Channel::Rx0, false);
-        let _ = self.dev.borrow().set_enable_module(Channel::Rx1, false);
-    }
-}
-
-// Tx Streamers
-
-pub struct TxSyncStream<'a, T: SampleFormat, D: BladeRF> {
-    pub(crate) dev: &'a D,
-    pub(crate) layout: ChannelLayoutTx,
-    pub(crate) _format: PhantomData<T>,
-}
-
-impl<'a, T: SampleFormat> TxSyncStream<'a, T, BladeRf1> {
-    pub fn write(&self, buffer: &[T], timeout: Duration) -> Result<()> {
-        let res = unsafe {
-            sys::bladerf_sync_tx(
-                self.dev.device,
-                buffer.as_ptr() as *const _,
-                buffer.len() as u32,
-                std::ptr::null_mut(),
-                timeout.as_millis() as u32,
-            )
-        };
-        check_res!(res);
-        Ok(())
-    }
-
-    pub fn reconfigure<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-    ) -> Result<TxSyncStream<'a, NF, BladeRf1>> {
-        unsafe {
-            self.dev
-                .set_sync_config::<NF>(config, ChannelLayout::TxSISO)?;
-        }
-
-        Ok(TxSyncStream {
-            dev: self.dev,
-            layout: ChannelLayoutTx::SISO(crate::TxChannel::Tx0),
-            _format: PhantomData,
-        })
-    }
-
-    pub fn enable(&self) -> Result<()> {
-        self.dev.set_enable_module(Channel::Tx0, true)
-    }
-
-    pub fn disable(&self) -> Result<()> {
-        self.dev.set_enable_module(Channel::Tx0, false)
-    }
-}
-
-impl<'a, T: SampleFormat> TxSyncStream<'a, T, BladeRf2> {
-    pub fn write(&self, buffer: &[T], timeout: Duration) -> Result<()> {
-        let res = unsafe {
-            sys::bladerf_sync_tx(
-                self.dev.device,
-                buffer.as_ptr() as *const _,
-                buffer.len() as u32,
-                std::ptr::null_mut(),
-                timeout.as_millis() as u32,
-            )
-        };
-        check_res!(res);
-        Ok(())
-    }
-
-    pub fn reconfigure<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-        layout: ChannelLayoutTx,
-    ) -> Result<TxSyncStream<'a, NF, BladeRf2>> {
-        unsafe {
-            self.dev.set_sync_config::<NF>(config, layout.into())?;
-        }
-
-        Ok(TxSyncStream {
-            dev: self.dev,
-            layout,
-            _format: PhantomData,
-        })
-    }
-
-    pub fn enable(&self) -> Result<()> {
-        match self.layout {
-            ChannelLayoutTx::SISO(ch) => self.dev.set_enable_module(ch.into(), true),
-            ChannelLayoutTx::MIMO => {
-                self.dev.set_enable_module(Channel::Tx0, true)?;
-                self.dev.set_enable_module(Channel::Tx1, true)?;
-                Ok(())
-            }
-        }
-    }
-
-    pub fn disable(&self) -> Result<()> {
-        match self.layout {
-            ChannelLayoutTx::SISO(ch) => self.dev.set_enable_module(ch.into(), false),
-            ChannelLayoutTx::MIMO => {
-                self.dev.set_enable_module(Channel::Tx0, false)?;
-                self.dev.set_enable_module(Channel::Tx1, false)?;
-                Ok(())
-            }
-        }
-    }
-}
-
-impl<'a, T: SampleFormat> TxSyncStream<'a, T, BladeRfAny> {
-    pub fn write(&self, buffer: &[T], timeout: Duration) -> Result<()> {
-        let res = unsafe {
-            sys::bladerf_sync_tx(
-                self.dev.device,
-                buffer.as_ptr() as *const _,
-                buffer.len() as u32,
-                std::ptr::null_mut(),
-                timeout.as_millis() as u32,
-            )
-        };
-        check_res!(res);
-        Ok(())
-    }
-
-    pub fn reconfigure<NF: SampleFormat>(
-        self,
-        config: &SyncConfig,
-        layout: ChannelLayoutTx,
-    ) -> Result<TxSyncStream<'a, NF, BladeRfAny>> {
-        unsafe {
-            self.dev.set_sync_config::<NF>(config, layout.into())?;
-        }
-
-        Ok(TxSyncStream {
-            dev: self.dev,
-            layout,
-            _format: PhantomData,
-        })
-    }
-
-    pub fn enable(&self) -> Result<()> {
-        match self.layout {
-            ChannelLayoutTx::SISO(ch) => self.dev.set_enable_module(ch.into(), true),
-            ChannelLayoutTx::MIMO => {
-                self.dev.set_enable_module(Channel::Tx0, true)?;
-                self.dev.set_enable_module(Channel::Tx1, true)?;
-                Ok(())
-            }
-        }
-    }
-
-    pub fn disable(&self) -> Result<()> {
-        match self.layout {
-            ChannelLayoutTx::SISO(ch) => self.dev.set_enable_module(ch.into(), false),
-            ChannelLayoutTx::MIMO => {
-                self.dev.set_enable_module(Channel::Tx0, false)?;
-                self.dev.set_enable_module(Channel::Tx1, false)?;
-                Ok(())
-            }
-        }
-    }
-}
-
-impl<T: SampleFormat, D: BladeRF> Drop for TxSyncStream<'_, T, D> {
-    fn drop(&mut self) {
-        // Ignore the results, just try disable both channels even if they don't exist on the dev.
-        let _ = self.dev.set_enable_module(Channel::Tx0, false);
-        let _ = self.dev.set_enable_module(Channel::Tx1, false);
     }
 }
